move gist && copyOver to files-tab.js
