+++ src/app/files-tab.js
@@ -1,4 +1,8 @@
+/* global alert, confirm, prompt */
 var yo = require('yo-yo')
+var $ = require('jquery')
+var QueryParams = require('./query-params')
+var queryParams = new QueryParams()
 
 // -------------- styling ----------------------
 var csjs = require('csjs-inject')
@@ -42,5 +46,55 @@ function filesTab (container, appAPI, events, opts) {
       </div>
     </div>
   `
+  // ------------------ gist publish --------------
+
+  el.querySelector('#gist').addEventListener('click', function () {
+    if (confirm('Are you sure you want to publish all your files anonymously as a public gist on github.com?')) {
+      appAPI.packageFiles((error, packaged) => {
+        if (error) {
+          console.log(error)
+        } else {
+          var description = 'Created using browser-solidity: Realtime Ethereum Contract Compiler and Runtime. \n Load this file by pasting this gists URL or ID at https://ethereum.github.io/browser-solidity/#version=' + queryParams.get().version + '&optimize=' + queryParams.get().optimize + '&gist='
+          console.log(packaged)
+          $.ajax({
+            url: 'https://api.github.com/gists',
+            type: 'POST',
+            data: JSON.stringify({
+              description: description,
+              public: true,
+              files: packaged
+            })
+          }).done(function (response) {
+            if (response.html_url && confirm('Created a gist at ' + response.html_url + ' Would you like to open it in a new window?')) {
+              window.open(response.html_url, '_blank')
+            }
+          }).fail(function (xhr, text, err) {
+            console.log('fail', text)
+            alert('Failed to create gist: ' + (err || 'Unknown transport error'))
+          })
+        }
+      })
+    }
+  })
+  el.querySelector('#copyOver').addEventListener('click', function () {
+    var target = prompt(
+      'To which other browser-solidity instance do you want to copy over all files?',
+      'https://ethereum.github.io/browser-solidity/'
+    )
+    if (target === null) {
+      return
+    }
+    appAPI.packageFiles((error, packaged) => {
+      if (error) {
+        console.log(error)
+      } else {
+        $('<iframe/>', {
+          src: target,
+          style: 'display:none;',
+          load: function () { this.contentWindow.postMessage(['loadFiles', packaged], '*') }
+        }).appendTo('body')
+      }
+    })
+  })
   container.appendChild(el)
 }
